refactor: rest docs 패키지 이동
