+++ src/test/java/project/volunteer/document/ScheduleControllerTest.java
@@ -1,0 +1,4 @@
+package project.volunteer.document;
+
+public class ScheduleControllerTest {
+}
