+++ src/test/java/project/volunteer/document/restdocs/EnumViewController.java
@@ -1,4 +1,4 @@
-package project.volunteer.restdocs.document;
+package project.volunteer.document.restdocs;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,8 +11,8 @@ import project.volunteer.domain.recruitment.domain.Week;
 import project.volunteer.global.common.component.HourFormat;
 import project.volunteer.global.common.converter.CodeCommonType;
 import project.volunteer.global.common.dto.StateResponse;
-import project.volunteer.restdocs.document.dto.APIResponseDto;
-import project.volunteer.restdocs.document.dto.EnumDocs;
+import project.volunteer.document.restdocs.dto.APIResponseDto;
+import project.volunteer.document.restdocs.dto.EnumDocs;
 
 import java.util.Arrays;
 import java.util.Map;
