+++ src/test/java/project/volunteer/document/restdocs/util/DocumentFormatGenerator.java
@@ -1,4 +1,4 @@
-package project.volunteer.restdocs.document.util;
+package project.volunteer.document.restdocs.util;
 
 import org.springframework.restdocs.snippet.Attributes;
