+++ src/test/java/project/volunteer/domain/notice/api/NoticeCommentControllerTest.java
@@ -36,7 +36,7 @@ import project.volunteer.domain.user.domain.Role;
 import project.volunteer.domain.user.domain.User;
 import project.volunteer.global.common.component.*;
 import project.volunteer.global.common.dto.CommentContentParam;
-import project.volunteer.restdocs.document.config.RestDocsConfiguration;
+import project.volunteer.document.restdocs.config.RestDocsConfiguration;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
