+++ src/test/java/project/volunteer/domain/recruitment/api/RecruitmentEditControllerTest.java
@@ -36,7 +36,7 @@ import project.volunteer.domain.user.domain.Role;
 import project.volunteer.domain.user.domain.User;
 import project.volunteer.global.infra.s3.FileService;
 import project.volunteer.global.test.WithMockCustomUser;
-import project.volunteer.restdocs.document.config.RestDocsConfiguration;
+import project.volunteer.document.restdocs.config.RestDocsConfiguration;
 
 import java.io.FileInputStream;
 import java.io.IOException;
